Move legacy form view to utilities, separate api requests
