+++ client/galaxy/scripts/components/Workflow/icons.js
