+++ client/galaxy/scripts/qunit/tests/workflow_editor_tests.js
@@ -1,18 +1,14 @@
 /* global QUnit */
 import $ from "jquery";
-import testApp from "qunit/test-app";
+import DatatypesMapping from "qunit/test-data/json/datatypes.mapping.json";
 import sinon from "sinon";
 import Utils from "utils/utils";
-import { WorkflowView } from "mvc/workflow/workflow-view";
+import WorkflowManager from "mvc/workflow/workflow-manager";
 import { Node } from "mvc/workflow/workflow-node";
 import { NodeView } from "mvc/workflow/workflow-view-node";
 import Terminals from "mvc/workflow/workflow-terminals";
 import TerminalsView from "mvc/workflow/workflow-view-terminals";
 import Connector from "mvc/workflow/workflow-connector";
-import { getAppRoot } from "onload/loadConfig";
-
-// window.show_modal = function(a, b, c) {};
-// window.hide_modal = function() {};
 
 // create body and app
 var create_app = function() {
@@ -28,24 +24,24 @@ var create_app = function() {
     );
 
     // build app
-    return new WorkflowView({
-        id: null,
-        urls: { get_datatypes: getAppRoot() + "api/datatypes/mapping" },
-        workflows: []
-    });
+    return new WorkflowManager(
+        {
+            datatypes: [],
+            datatypes_mapping: DatatypesMapping
+        },
+        $("#canvas-container")
+    );
 };
 
 QUnit.module("Input terminal model test", {
     beforeEach: function() {
-        testApp.create();
         this.app = create_app();
-        this.node = new Node(this.app.workflow, {});
+        this.node = new Node(this.app, {});
         this.input = { extensions: ["txt"], multiple: false, optional: false };
-        this.input_terminal = new Terminals.InputTerminal({ app: this.app.workflow, input: this.input });
+        this.input_terminal = new Terminals.InputTerminal({ app: this.app, input: this.input });
         this.input_terminal.node = this.node;
     },
     afterEach: function() {
-        testApp.destroy();
         delete this.node;
     },
     multiple: function() {
@@ -286,15 +282,11 @@ QUnit.test("initial redraw", function(assert) {
 
 QUnit.module("Input collection terminal model test", {
     beforeEach: function() {
-        testApp.create();
         this.app = create_app();
-        this.node = new Node(this.app.workflow, {});
+        this.node = new Node(this.app, {});
         this.input = { extensions: ["txt"], collection_types: ["list"] };
-        this.input_terminal = new Terminals.InputCollectionTerminal({ app: this.app.workflow, input: this.input });
+        this.input_terminal = new Terminals.InputCollectionTerminal({ app: this.app, input: this.input });
         this.input_terminal.node = this.node;
-    },
-    afterEach: function() {
-        testApp.destroy();
     }
 });
 
@@ -334,24 +326,20 @@ QUnit.test("Collection output cannot connect to different collection input type"
 
 QUnit.module("Node unit test", {
     beforeEach: function() {
-        testApp.create();
         this.input_terminal = { destroy: sinon.spy(), redraw: sinon.spy() };
         this.output_terminal = { destroy: sinon.spy(), redraw: sinon.spy() };
         this.app = create_app();
-        this.node = this.app.workflow.prebuildNode("tool", "newnode");
+        this.node = this.app.prebuildNode("tool", "newnode");
         this.element = this.node.element;
         this.node.input_terminals.i1 = this.input_terminal;
         this.node.output_terminals.o1 = this.output_terminal;
-    },
-    afterEach: function() {
-        testApp.destroy();
     },
     $: function(selector) {
         return $(this.node.element.find(selector));
     },
     expect_workflow_node_changed: function(assert, f) {
         const node = this.node;
-        const node_changed_spy = sinon.spy(this.app.workflow, "node_changed");
+        const node_changed_spy = sinon.spy(this.app, "node_changed");
         f();
         assert.ok(node_changed_spy.calledWith(node));
     },
@@ -384,7 +372,7 @@ QUnit.test("make active", function(assert) {
 });
 
 QUnit.test("destroy", function(assert) {
-    var remove_node_spy = sinon.spy(this.app.workflow, "remove_node");
+    var remove_node_spy = sinon.spy(this.app, "remove_node");
     this.node.destroy();
     assert.ok(this.input_terminal.destroy.called);
     assert.ok(this.output_terminal.destroy.called);
@@ -521,24 +509,19 @@ QUnit.test("update_field_data destroys old terminals", function(assert) {
 
 QUnit.module("create_node", {
     beforeEach: function() {
-        testApp.create();
         this.app = create_app();
-    },
-    afterEach: function() {
-        testApp.destroy();
     }
 });
 
 QUnit.test("node added to workflow", function(assert) {
-    var add_node_spy = sinon.spy(this.app.workflow, "add_node");
-    var node = this.app.workflow.create_node("tool", "Cat Files", "cat1");
+    var add_node_spy = sinon.spy(this.app, "add_node");
+    var node = this.app.create_node("tool", "Cat Files", "cat1");
     assert.ok(add_node_spy.calledWith(node));
 });
 
 // global NodeView
 QUnit.module("Node view ", {
     beforeEach: function() {
-        testApp.create();
         this.app = create_app();
         this.set_for_node({
             input_terminals: {},
@@ -554,7 +537,7 @@ QUnit.module("Node view ", {
     },
     set_for_node: function(node) {
         var element = $("<div><div class='toolFormBody'></div></div>");
-        this.view = new NodeView(this.app.workflow, { node: node, $el: element });
+        this.view = new NodeView(this.app, { node: node, $el: element });
     },
     connectAttachedTerminal: function(inputType, outputType) {
         this.view.addDataInput({ name: "TestName", extensions: [inputType] });
@@ -676,7 +659,7 @@ QUnit.test("replacing terminal on data input destroys invalid connections", func
     var connector_destroy_spy = sinon.spy(connector, "destroy");
     // Replacing input with same name - but now of type bam should destroy connection.
     this.view.addDataInput({ name: "TestName", extensions: ["bam"] }, newElement);
-    var terminal = newElement.find(".input-terminal")[0].terminal;
+    newElement.find(".input-terminal")[0].terminal;
     assert.ok(connector_destroy_spy.called);
 });
 
@@ -687,7 +670,7 @@ QUnit.test("replacing terminal on data input with collection changes mapping vie
     this.view.addDataInput({ name: "TestName", extensions: ["txt"], input_type: "dataset_collection" }, newElement);
     // Input type changed to dataset_collection - old connections are reset.
     // Would be nice to preserve these connections and make them map over.
-    var terminal = newElement.find(".input-terminal")[0].terminal;
+    newElement.find(".input-terminal")[0].terminal;
     assert.ok(connector_destroy_spy.called);
 });
 
@@ -696,18 +679,17 @@ QUnit.test("replacing terminal on data collection input with simple input change
     var newElement = $("<div class='inputs'></div>");
     var connector_destroy_spy = sinon.spy(connector, "destroy");
     this.view.addDataInput({ name: "TestName", extensions: ["txt"], input_type: "dataset" }, newElement);
-    var terminal = newElement.find(".input-terminal")[0].terminal;
+    newElement.find(".input-terminal")[0].terminal;
     assert.ok(connector_destroy_spy.called);
 });
 
 // global InputTerminalView
 QUnit.module("Input terminal view", {
     beforeEach: function() {
-        testApp.create();
         this.app = create_app();
         this.node = { input_terminals: [] };
         this.input = { name: "i1", extensions: "txt", multiple: false };
-        this.view = new TerminalsView.InputTerminalView(this.app.workflow, {
+        this.view = new TerminalsView.InputTerminalView(this.app, {
             node: this.node,
             input: this.input
         });
@@ -728,11 +710,10 @@ QUnit.test("terminal element", function(assert) {
 
 QUnit.module("Output terminal view", {
     beforeEach: function() {
-        testApp.create();
         this.app = create_app();
         this.node = { output_terminals: [] };
         this.output = { name: "o1", extensions: "txt" };
-        this.view = new TerminalsView.OutputTerminalView(this.app.workflow, {
+        this.view = new TerminalsView.OutputTerminalView(this.app, {
             node: this.node,
             output: this.output
         });
@@ -820,15 +801,6 @@ QUnit.test("equal", function(assert) {
     assert.ok(!this.listType().equal(Terminals.NULL_COLLECTION_TYPE_DESCRIPTION));
 });
 
-QUnit.module("TerminalMapping", {
-    beforeEach: function() {
-        testApp.create();
-    },
-    afterEach: function() {
-        testApp.destroy();
-    }
-});
-
 QUnit.test("default constructor", function(assert) {
     var terminal = {};
     var mapping = new Terminals.TerminalMapping({ terminal: terminal });
@@ -859,15 +831,6 @@ QUnit.test("disableMapOver", function(assert) {
 });
 
 QUnit.module("terminal mapping logic", {
-    beforeEach: function() {
-        testApp.create();
-    },
-    afterEach: function() {
-        testApp.destroy();
-        if (this.inputTerminal1) {
-            this.inputTerminal1.element.remove();
-        }
-    },
     newInputTerminal: function(mapOver, input, node) {
         input = input || {};
         node = node || this.newNode();
@@ -876,8 +839,8 @@ QUnit.module("terminal mapping logic", {
         }
         var inputEl = $("<div>")[0];
         const app = create_app();
-        var inputTerminal = new Terminals.InputTerminal({ app: app.workflow, element: inputEl, input: input });
-        var inputTerminalMapping = new Terminals.TerminalMapping({ terminal: inputTerminal });
+        var inputTerminal = new Terminals.InputTerminal({ app: app, element: inputEl, input: input });
+        new Terminals.TerminalMapping({ terminal: inputTerminal });
         inputTerminal.node = node;
         if (mapOver) {
             inputTerminal.setMapOver(new Terminals.CollectionTypeDescription(mapOver));
@@ -893,7 +856,7 @@ QUnit.module("terminal mapping logic", {
         const inputEl = $("<div>")[0];
         const app = create_app();
         const inputTerminal = new Terminals.InputCollectionTerminal({
-            app: app.workflow,
+            app: app,
             element: inputEl,
             input: input
         });
@@ -929,7 +892,7 @@ QUnit.module("terminal mapping logic", {
             datatypes: output.extensions,
             collection_type: collectionType
         });
-        var outputTerminalMapping = new Terminals.TerminalMapping({ terminal: outputTerminal });
+        new Terminals.TerminalMapping({ terminal: outputTerminal });
         outputTerminal.node = node;
         if (mapOver) {
             outputTerminal.setMapOver(new Terminals.CollectionTypeDescription(mapOver));
@@ -939,7 +902,7 @@ QUnit.module("terminal mapping logic", {
     newNode: function() {
         var nodeEl = $("<div>")[0];
         var app = create_app();
-        return new Node(app.workflow, { element: nodeEl });
+        return new Node(app, { element: nodeEl });
     },
     _addExistingOutput: function(terminal, output, connected) {
         var self = this;
@@ -1027,7 +990,7 @@ QUnit.test("unconstrained input can be mapped over", function(assert) {
 
 QUnit.test("unmapped input can be mapped over if matching connected input terminals map type", function(assert) {
     this.inputTerminal1 = this.newInputTerminal();
-    var connectedInput1 = this.addConnectedInput(this.inputTerminal1);
+    this.addConnectedInput(this.inputTerminal1);
     var connectedInput2 = this.addConnectedInput(this.inputTerminal1);
     connectedInput2.setMapOver(new Terminals.CollectionTypeDescription("list"));
     this.verifyAttachable(assert, this.inputTerminal1, "list");
@@ -1044,7 +1007,7 @@ QUnit.test(
     "unmapped input can be attached to by output collection if matching connected input terminals map type",
     function(assert) {
         this.inputTerminal1 = this.newInputTerminal();
-        var connectedInput1 = this.addConnectedInput(this.inputTerminal1);
+        this.addConnectedInput(this.inputTerminal1);
         var connectedInput2 = this.addConnectedInput(this.inputTerminal1);
         connectedInput2.setMapOver(new Terminals.CollectionTypeDescription("list"));
         var outputTerminal = this.newOutputCollectionTerminal("list");
@@ -1056,7 +1019,7 @@ QUnit.test(
     "unmapped input cannot be attached to by output collection if matching connected input terminals don't match map type",
     function(assert) {
         this.inputTerminal1 = this.newInputTerminal();
-        var connectedInput1 = this.addConnectedInput(this.inputTerminal1);
+        this.addConnectedInput(this.inputTerminal1);
         var connectedInput2 = this.addConnectedInput(this.inputTerminal1);
         connectedInput2.setMapOver(new Terminals.CollectionTypeDescription("list"));
         var outputTerminal = this.newOutputCollectionTerminal("paired");
@@ -1068,7 +1031,7 @@ QUnit.test(
     "unmapped input can be attached to by output collection if effective output type (output+mapover) is same as mapped over input",
     function(assert) {
         this.inputTerminal1 = this.newInputTerminal();
-        var connectedInput1 = this.addConnectedInput(this.inputTerminal1);
+        this.addConnectedInput(this.inputTerminal1);
         var connectedInput2 = this.addConnectedInput(this.inputTerminal1);
         connectedInput2.setMapOver(new Terminals.CollectionTypeDescription("list:paired"));
         var outputTerminal = this.newOutputCollectionTerminal("paired");
@@ -1081,7 +1044,7 @@ QUnit.test(
     "unmapped input cannot be attached to by output collection if effective output type (output+mapover) is not same as mapped over input (1)",
     function(assert) {
         this.inputTerminal1 = this.newInputTerminal();
-        var connectedInput1 = this.addConnectedInput(this.inputTerminal1);
+        this.addConnectedInput(this.inputTerminal1);
         var connectedInput2 = this.addConnectedInput(this.inputTerminal1);
         connectedInput2.setMapOver(new Terminals.CollectionTypeDescription("list:paired"));
         var outputTerminal = this.newOutputCollectionTerminal("list");
@@ -1094,7 +1057,7 @@ QUnit.test(
     "unmapped input cannot be attached to by output collection if effective output type (output+mapover) is not same as mapped over input (2)",
     function(assert) {
         this.inputTerminal1 = this.newInputTerminal();
-        var connectedInput1 = this.addConnectedInput(this.inputTerminal1);
+        this.addConnectedInput(this.inputTerminal1);
         var connectedInput2 = this.addConnectedInput(this.inputTerminal1);
         connectedInput2.setMapOver(new Terminals.CollectionTypeDescription("list:paired"));
         var outputTerminal = this.newOutputCollectionTerminal("list");
